// src/interfaces/IRequestCoverage/DocumentData.ts
export interface DocumentData {
    nombre:            string;
    fechaNacimiento:   string;
    direccion:         string;
    nroDireccion:      number;
    depto:             number;
    ciudad:            string;
    comuna:            string;
    telefono:          number;
    celular:           number;
    correoElectronico: string;
}